Add many features to alcotest_ext and use it (#9434)
This is mostly new code that is not exercised in production, only in
tests. The alcotest_ext library now has a lot of features and is usable
for our project. Many more options and small features will probably be
added eventually when people use it outside of semgrep but for us, it's
pretty solid already.

User interface
--

The generated command is an executable with 3 subcommands: `run`
(default), `status`, and `approve`. It must be called from the same
location everytime because it uses files to store test results and
expectations in `_test_results`. The test program is invoked with
`./test` at the root of the project. `./test --help` provides user
documentation:


![image](https://github.com/semgrep/semgrep/assets/343265/3563601a-6e59-4592-9220-06258a374073)

File storage
--

Each test is identified by a unique 12-character hexadecimal hash that
will work nicely for up to at least a million tests. This ID is more
compact and generally easier to deal with than the full test name from
which it derives. So, the test program stores two kinds of test data:
* `status/`: folder containing the latest results obtained for the
tests, including the outcome of the last run (success/failure) and any
captured output.
* `expect/`: folder containing the expected output for tests that
capture output. When a test is approved using the `approve` subcommand,
the output from the last run is moved there.

The expected output is kept under version control while the status
folder is not.

The storage folder looks like this for tiny project (alcotest_ext's own
tests):


![image](https://github.com/semgrep/semgrep/assets/343265/0d20c12a-f8c4-465c-8b4a-03e471fd0d37)

Usage in our project
--

The `./test status` lists all the details about each selected test
without re-running them. Use `-s <TEXT>` to select a test by ID or by a
substring contained in the full test name. Eventually, there will be
more filters such as a way to filter tests by tag. Right now, tags are
displayed by they can't be used to filters tests on the command line
yet.


![image](https://github.com/semgrep/semgrep/assets/343265/6293f5d1-dc1e-4d47-8e3c-1d9d41f823cb)

After running tests with `make test` or directly with `./test`, the
tests that need fixing or approval are listed in a "short status"
section at the end with some stats. The full details about all the tests
are shown in the "long status" that precedes.

When working on fixing one or a few tests, run `make retest`. This
invokes `./test --lazy` which only runs the tests that previously
failed.

Plans for the future
--

1. Use this tool and polish it up.
2. Figure out what to do with Lwt because I'm not sure we need a
separate interface for it - just have each test run the Lwt event loop,
no? I'd rather avoid having to use functors or code duplication if we
can avoid it as it makes the source harder to follow and extend.
3. Figure out a path to make this library usable outside the company.
Vague options include: keep it as a layer on top of Alcotest as it is
today, merge it with the Alcotest code base, or both.