Alcotest_ext improvements (#9528)
Major changes:
* Alcotest_ext no longer uses Alcotest to run the tests.
* The parts that depend on Lwt moved to a separate library and package.
The same is easy to do for Jane Street's `Async`.
* The output was improved and I believe is less confusing than before.
It's not as beautiful as Alcotest but it shows a lot more info.

Alcotest is still recommended for writing assertions (e.g.
`Alcotest.(check int) "matches" 42 res`) which raise exceptions
signaling test failure and print good error messages.

Some environments such as JavaScript require a test to return a promise,
unlike pure Lwt which allows a test to be made synchronous with
`Lwt_main.run` or equivalent. This is why we have offer the option to
define tests that return promises (deferred computations). This is made
possible by providing a monad-like record (module `Alcotest_ext.Mona`)
rather than via a functor which I think would have made the library code
more contrived.

Next steps:

1. Rename: Alcotest_ext -> Testo.
2. Move out of the semgrep code base.
3. Publish and announce an opam package.
4. Dogfooding, interface stabilization, documentation.

Screenshot 1 shows 3 failing tests whose details are being printed at
the end of the run:

![image](https://github.com/semgrep/semgrep/assets/343265/05316c4e-fdc0-4216-93f8-23a79d25333b)

Screenshot 2 shows the output for a successful run of another suite:

![image](https://github.com/semgrep/semgrep/assets/343265/c1696cf6-a4b5-4fc7-b330-f291d0ef3c28)

Screenshot 3 shows the beginning of a run:

![image](https://github.com/semgrep/semgrep/assets/343265/9418f114-fb79-44aa-aa72-0116a6c79e2c)