fix: masking on macos
MacOS ridiculously symlinks their `$TMPDIR =
/var/folders/lz/<somestring>/T/`
to `/private/var/folders/...`. This means that if for whatever reason
someone prints the real path of a tmp folder instead of the symlinked
path, TestO wouldnt mask it.

Also, $TMPDIR on MacOS has a trailing slash for some reason, when on
Linux `Filename.get_temp_dir_name` does not which also broke some things