Fix "Bad file descriptor" errors on Windows
Previously, we dealt with stdxxx with nested callbacks that redirected
stdout to a file, then redirected stderr to stdout, in separate layers.
While this apparently worked fine on proper POSIX systems, on Windows,
this would produce a `Sys_error("Bad file descriptor")` due to the
redirected file descriptors being closed. The fix is to generalize from

    with_redirect_fd : from:file_descr -> to_:file_descr -> (unit -> 'a) -> 'a

to

    `with_redirect_fds : from:file_descr list -> to_:file_descr -> (unit -> 'a) -> 'a`

where the latter takes a list of file descriptors and redirects all of
them to the target, allowing us to call `with_redirect_fds
~from:[Unix.stderr; Unix.stdout] ~to_:fd func` to have both stderr and
stdout redirected to `fd` while computing `func`, without complicating
the redirects by trying to first run stderr thru stdout.