Make testo Windows compatible (#121)
* Add missing expected test output

* Run test commands in bash

On Windows `Sys.command` runts commands thru cmd.exe, and it is
infeasible to get the same output for these kinds of textual expectation
tests running in such a different environment.  With this change, we
instead run all commands via bash, which gives us a reliable posix
environment.

* Fix "Bad file descriptor" errors on Windows

Previously, we dealt with stdxxx with nested callbacks that redirected
stdout to a file, then redirected stderr to stdout, in separate layers.
While this apparently worked fine on proper POSIX systems, on Windows,
this would produce a `Sys_error("Bad file descriptor")` due to the
redirected file descriptors being closed. The fix is to generalize from

    with_redirect_fd : from:file_descr -> to_:file_descr -> (unit -> 'a) -> 'a

to

    `with_redirect_fds : from:file_descr list -> to_:file_descr -> (unit -> 'a) -> 'a`

where the latter takes a list of file descriptors and redirects all of
them to the target, allowing us to call `with_redirect_fds
~from:[Unix.stderr; Unix.stdout] ~to_:fd func` to have both stderr and
stdout redirected to `fd` while computing `func`, without complicating
the redirects by trying to first run stderr thru stdout.

* Set stdout and stderr to binary mode

* Fix empty env var check on Windows

On Windows `Unix.putenv "B" ""; Sys.getenv_opt "B" = None`

* Remove windows_todo

We can use skips instead, and `Sys.win32` is a better way to test for
whether we are on Windows.

* Skip tests failing in Windows

These will be fixed in a followup, on top of a passing CI

* Fix path masking in Windows

* Dont run the tests for multiple jobs on Windows

* Handle Windows paths in man page

* Explain why we posixify masked paths

* Update changelog