Rename Worker module -> Multiprocess (#128)
This PR changes the name of one module and adds two lines of comments in Multiprocess.ml.

This is an internal change only. It's not part of official API even though HTML documentation is published.

The number of lines increased because the HTML/CSS docs were regenerated using a newer version of odoc.