Check for both files before attempting a diff (#155)
Don't fail when printing a test status if the test didn't stash an output file like it should.

Test plan: this is dependent on an internal malfunction that is hard to reproduce and was causing a crash of the test suite. To reproduce, insert a `Sys.remove !!path_to_output;`​ at the beginning of the `show_diff`​ function. `show_diff`​ should no longer raise an exception (but of course will likely cause a failure later).

PR checklist:

- [x] Purpose of the code is evident to future readers
- [x] Tests are included or a PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was added to `CHANGES.md` for any user-facing change

Check out [`CONTRIBUTING.md`](https://github.com/semgrep/testo/blob/main/CONTRIBUTING.md) for more details.