Add support for tag selection with boolean queries (#157)
This implements support for tag selection such as `./test -t '(foo or bar) and not e2e'`​ or just `./test 'not e2e'`​ which is what I'm going to use it for soon.

I also added a `show-tags`​ subcommand to list the available tags.

Closes https://github.com/semgrep/testo/issues/5

Minor note: The HTML help page for the new `show-tags`​ subcommand is wrong because it uses the main branch of the testo repo to generate a demo executable. I'll regenerate it as soon as this PR is merged.

PR checklist:

- [x] Purpose of the code is evident to future readers
- [x] Tests are included or a PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was added to `CHANGES.md` for any user-facing change

Check out [`CONTRIBUTING.md`](https://github.com/semgrep/testo/blob/main/CONTRIBUTING.md) for more details.